Update sources for Draftail 0.17 compatibility
It would be too much efforts to support sources for Draftail 0.17 (as of Wagtail 2.0) and Draftail 0.7 (as of WagtailDraftail 0.7), therefore taking the highest of the two.

WagtailDraftail will have to receive an update to keep compatibility with this package which is a pain but at the same time, it's pretty unlikely that WagtailDraftail will be used alongside with Wagtail 2.0 now that Draftail is integreated within Wagtail.