BIG: DB convs approach
You need a strong background in draftail customization to understand this

We're stuck with the following bad things:
* We gotta save to the DB in the HTML format we're going to render.
    This means draftail is sharing that format so we have to save
    everything draftail needs.
* There needs to be a system-unique DOM selector rule to find tags in
    the HTML to convert to contentstate for the draftail editor.

To solve both at the same time we add an attr called modelchooser_choice_{}
which holds the choice id, and the name of the attr resolves the unique
DOM selector thing, because we can use span[modelchooser_choice_{}].

That's fine, frontend can just ignore it, but the big drawback is that
draftail needs the DOM element's children so it can display the block,
but then the frontend gets it too.  Frontend does the correct thing and
replaces the text with live data, but you see a flicker of the alt
text.  Really annoying, but will resolve soon with some kind of
rewriter solution.